I wrote a lot of thing, but i don't know what i doing, made some payment related thing and i forgot

+++ api/Models/Entity/NormalDB/ParkingLots.cs
@@ -29,6 +29,6 @@ namespace api.Models.Entity.NormalDB
     public class LotPrices
     {
         public required string time { get; set; }
-        public required double price { get; set; }
+        public required decimal price { get; set; }
     }
 }
